Separate libnvram_list tests
